+++ src/hooks/useRequests.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import dayjs from "dayjs";
-import axiosIns from "./AxiosInstance";
+import axiosIns from "../utils/AxiosInstance";
 
 const useRequests = (setLoading, setCalls, setHasNextPage) => {
   const { REACT_APP_API_BASE_URL, REACT_APP_USERNAME, REACT_APP_PASSWORD } =
